Small fixes

+++ BilliardServer.DataAccess/Entities/UserEntity.cs
@@ -2,7 +2,7 @@
 
 namespace BilliardServer.Infrastructure.Entities
 {
-    public class UserEntity : IdentityUser<int>
+    public class UserEntity : IdentityUser<long>
     {
         public string Name { get; set; } = string.Empty;
         public int Avatar { get; set; }
